Add `Inigo.dhall` capability
This adds the ability to configure packages with an `Inigo.dhall` file.
If Inigo doesn't find an `Inigo.toml` file, it will check for an
`Inigo.dhall` file. If it finds both or neither, then it errors.

The format of the `Inigo.dhall` file is described in
`docs/src/dhall.rst`.